email integration added

+++ src/main/java/com/cornershop/ecommerce/service/CategoryService.java
@@ -1,12 +1,16 @@
 package com.cornershop.ecommerce.service;
 
+import com.cornershop.ecommerce.exception.CategoryDeleteException;
 import com.cornershop.ecommerce.exception.CategoryDuplicateException;
 import com.cornershop.ecommerce.exception.CategoryNotFoundException;
 import com.cornershop.ecommerce.model.Category;
 import com.cornershop.ecommerce.repository.CategoryRepository;
+
+import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
 
+import com.cornershop.ecommerce.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +19,9 @@ public class CategoryService {
 
     @Autowired
     private CategoryRepository categoryRepository;
+
+    @Autowired
+    private ProductRepository productRepository;
 
     public Category createCategory(Category category) {
         //categoryRepositoryde aynı category'yi html sayfasından girilirse eklememesi ve hata atması için
@@ -26,6 +33,10 @@ public class CategoryService {
     }
 
     public void deleteCategory(Long id) {
+        Long productCountOfCategory = productRepository.getProductCountOfCategoryId(id);
+        if(productCountOfCategory > 0){       //html sayfasında silmek istediğin category'ye ait ürün varsa ve 1 taneyse silmesin
+            throw  new CategoryDeleteException("You can not delete this category because category has" + productCountOfCategory + "products");
+        }
         categoryRepository.deleteById(id);
     }
 
@@ -34,9 +45,13 @@ public class CategoryService {
     }
 
     public List<Category> getAllCategoryList() {
-        return categoryRepository.findAll();
+        //category'leri id ye göre sıralamak için
+        List <Category> categoryList = categoryRepository.findAll();
+        categoryList.sort(Comparator.comparing(Category::getId));
+        return categoryList;
     }
     public Category updateCategory(Category category){
+
         return categoryRepository.save(category);
     }
 }
